Refactor scheduler to support minute and hour tasks
Split scheduler into separate minute and hour schedulers, moving scheduled pairing logic to a new pairing.go file. Updated bot and tests to use StartMinuteScheduler and StopMinuteScheduler. This improves clarity and allows for more flexible scheduling of background tasks.
