+++ internal/scheduler/scheduler.go
@@ -1,7 +1,6 @@
 package scheduler
 
 import (
-	"fmt"
 	"gamerpal/internal/config"
 	"gamerpal/internal/database"
 	"gamerpal/internal/pairing"
@@ -17,7 +16,8 @@ type Scheduler struct {
 	db             *database.DB
 	pairingService *pairing.PairingService
 	ticker         *time.Ticker
-	stopCh         chan struct{}
+	minuteStopCh   chan struct{}
+	hourStopCh     chan struct{}
 }
 
 // NewScheduler creates a new scheduler instance
@@ -27,98 +27,64 @@ func NewScheduler(session *discordgo.Session, cfg *config.Config, db *database.D
 		config:         cfg,
 		db:             db,
 		pairingService: pairingService,
-		stopCh:         make(chan struct{}),
+		minuteStopCh:   make(chan struct{}),
+		hourStopCh:     make(chan struct{}),
 	}
 }
 
-// Start begins the scheduler's background operations
-func (s *Scheduler) Start() {
+// StartMinuteScheduler starts a scheduler that runs every minute
+func (s *Scheduler) StartMinuteScheduler() {
 	// Check for scheduled pairings every minute
 	s.ticker = time.NewTicker(time.Minute)
 
 	go func() {
-		s.config.Logger.Info("Scheduler started - checking for scheduled pairings every minute")
+		s.config.Logger.Info("Minute scheduler started!")
 
 		for {
 			select {
 			case <-s.ticker.C:
 				s.checkAndExecuteScheduledPairings()
-			case <-s.stopCh:
-				s.config.Logger.Info("Scheduler stopping")
+			case <-s.minuteStopCh:
+				s.config.Logger.Info("Minute scheduler stopping")
 				return
 			}
 		}
 	}()
 }
 
-// Stop stops the scheduler
-func (s *Scheduler) Stop() {
+// StopMinuteScheduler stops the scheduler
+func (s *Scheduler) StopMinuteScheduler() {
 	if s.ticker != nil {
 		s.ticker.Stop()
 	}
-	close(s.stopCh)
+	close(s.minuteStopCh)
 }
 
-// checkAndExecuteScheduledPairings checks for due scheduled pairings and executes them
-func (s *Scheduler) checkAndExecuteScheduledPairings() {
-	scheduledPairings, err := s.db.GetScheduledPairings()
-	if err != nil {
-		s.config.Logger.Errorf("Error getting scheduled pairings: %v", err)
-		return
-	}
+func (s *Scheduler) StartHourScheduler() {
+	// Check for old log files every hour
+	s.ticker = time.NewTicker(time.Hour)
 
-	if len(scheduledPairings) == 0 {
-		return // No scheduled pairings
-	}
+	go func() {
+		s.config.Logger.Info("Hourly scheduler started!")
 
-	s.config.Logger.Infof("Found %d scheduled pairing(s) to execute", len(scheduledPairings))
-
-	for _, schedule := range scheduledPairings {
-		s.executeScheduledPairing(schedule)
-	}
+		for {
+			select {
+			case <-s.ticker.C:
+				if err := s.config.PruneOldLogFiles(); err != nil {
+					s.config.Logger.Errorf("Scheduler failed pruning old log files: %v", err)
+				}
+			case <-s.hourStopCh:
+				s.config.Logger.Info("Hourly scheduler stopping")
+				return
+			}
+		}
+	}()
 }
 
-// executeScheduledPairing executes a single scheduled pairing
-func (s *Scheduler) executeScheduledPairing(schedule database.RouletteSchedule) {
-	s.config.Logger.Infof("Executing scheduled pairing for guild %s (scheduled for %s)",
-		schedule.GuildID, schedule.ScheduledAt.Format("2006-01-02 15:04:05"))
-
-	// Execute pairing using the pairing service
-	result, err := s.pairingService.ExecutePairing(schedule.GuildID, false)
-	if err != nil {
-		s.config.Logger.Errorf("Error executing scheduled pairing for guild %s: %v", schedule.GuildID, err)
-		s.notifyFailedPairing(schedule.GuildID, fmt.Sprintf("Error executing pairing: %v", err))
-		return
+// StopHourScheduler stops the hourly scheduler
+func (s *Scheduler) StopHourScheduler() {
+	if s.ticker != nil {
+		s.ticker.Stop()
 	}
-
-	if !result.Success {
-		s.config.Logger.Errorf("Scheduled pairing failed for guild %s: %s", schedule.GuildID, result.ErrorMessage)
-		s.notifyFailedPairing(schedule.GuildID, result.ErrorMessage)
-	} else {
-		s.config.Logger.Infof("Successfully executed scheduled pairing for guild %s: %d pair(s) created",
-			schedule.GuildID, result.PairCount)
-
-		// Log the results
-		s.pairingService.LogPairingResults(schedule.GuildID, result, true)
-	}
-}
-
-// notifyFailedPairing sends a notification about failed automated pairing
-func (s *Scheduler) notifyFailedPairing(guildID, reason string) {
-	modLogChannelID := s.config.GetGamerPalsModActionLogChannelID()
-	if modLogChannelID == "" {
-		return
-	}
-
-	message := fmt.Sprintf("⚠️ **Automated Roulette Pairing Failed**\n\n"+
-		"**Guild:** %s\n"+
-		"**Reason:** %s\n"+
-		"**Time:** <t:%d:F>\n\n"+
-		"Please check the roulette system and manually trigger pairing if needed.",
-		guildID, reason, time.Now().Unix())
-
-	_, err := s.session.ChannelMessageSend(modLogChannelID, message)
-	if err != nil {
-		s.config.Logger.Errorf("Error sending failed pairing notification: %v", err)
-	}
+	close(s.hourStopCh)
 }
